Move auth helpers out of pages
