+++ frontend/src/components/auth/AuthErrorBanner.tsx
