+++ frontend/src/components/auth/AuthPageLayout.tsx
